Committ before trying to remove left and right
Manually revert a change I made in a previous commit where
I added flex boxes as the margins on either side of the
entire page. Seems like that was a bad idea since I need
different backgdound colors in the margins as I go down the
page. I remember now why coding is so tough hahahahahahaha.