Adjust dynamic margins for card section
Add dynamic margins as flexboxes to either side of the cardsection.
These margins needed to have a slightly smaller min width since the
horizontal space the cards take up is slightly more than that taken
up by the combined elements above. It is 50 pixels wider, so each
margin can be 25 pixels thinner than they can be in the above
sections.